CLI added progress bar for import command
