+++ ydb/public/lib/ydb_cli/commands/ydb_service_scripting.cpp
@@ -2,6 +2,7 @@
 
 #include <ydb/public/lib/ydb_cli/common/pretty_table.h>
 #include <ydb/public/lib/ydb_cli/common/query_stats.h>
+#include <ydb/public/lib/ydb_cli/common/interactive.h>
 
 #include <util/folder/path.h>
 #include <util/folder/dirut.h>
@@ -115,7 +116,7 @@ int TCommandExecuteYqlScript::Run(TConfig& config) {
                         paramBuilder->Build(),
                         FillSettings(settings)
                 );
-                
+
                 auto result = asyncResult.GetValueSync();
                 ThrowOnError(result);
                 PrintResponseHeader(result);
