+++ ydb/public/lib/ydb_cli/common/common.h
@@ -39,7 +39,6 @@ private:
 
 bool ReadFromFileIfExists(TString& filePath, const TString& fileName, TString& output, bool allowEmpty = false);
 TString ReadFromFile(TString& filePath, const TString& fileName, bool allowEmpty = false);
-size_t TermWidth();
 TString InputPassword();
 
 }
