+++ ydb/public/lib/ydb_cli/common/interactive.cpp
@@ -1,5 +1,13 @@
 #include "interactive.h"
 
+#if defined(_unix_)
+#include <sys/ioctl.h>
+#include <termios.h>
+
+#elif defined(_win_)
+#include <windows.h>
+#endif
+
 namespace NYdb {
 namespace NConsoleClient {
 
@@ -17,6 +25,42 @@ bool AskYesOrNo() {
         }
     }
     return false;
+}
+
+bool IsStdinInteractive() {
+#if defined(_win32_)
+    return _isatty(_fileno(stdin));
+#elif defined(_unix_)
+    return isatty(fileno(stdin));
+#endif
+    return true;
+}
+
+bool IsStdoutInteractive() {
+#if defined(_win32_)
+    return _isatty(_fileno(stdout));
+#elif defined(_unix_)
+    return isatty(fileno(stdout));
+#endif
+    return true;
+}
+
+std::optional<size_t> GetTerminalWidth() {
+    if (!IsStdoutInteractive())
+        return {};
+
+#if defined(_win32_)
+    CONSOLE_SCREEN_BUFFER_INFO screen_buf_info;
+    if (GetConsoleScreenBufferInfo(GetStdHandle(STD_OUTPUT_HANDLE), &screen_buf_info)) {
+        return screen_buf_info.srWindow.Right - screen_buf_info.srWindow.Left + 1;
+    }
+#elif defined(_unix_)
+    struct winsize size;
+    if (ioctl(STDOUT_FILENO, TIOCGWINSZ, &size) != -1) {
+        return size.ws_col;
+    }
+#endif
+    return {};
 }
 
 }
