+++ ydb/public/lib/ydb_cli/common/progress_bar.h
@@ -9,9 +9,13 @@ class TProgressBar {
 public:
     explicit TProgressBar(size_t capacity);
 
+    void SetProcess(size_t progress);
+
     void AddProgress(size_t value);
 
 private:
+    void Render();
+
     size_t Capacity = 0;
     size_t CurProgress = 0;
 };
